requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0
django-crispy-forms==1.13.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
